feat: Refactor Java client's search interface
