+++ clients/java/src/main/java/io/camunda/client/api/search/filter/DecisionRequirementsFilter.java
@@ -15,7 +15,7 @@
  */
 package io.camunda.client.api.search.filter;
 
-import io.camunda.client.api.search.request.TypedSearchRequest.SearchRequestFilter;
+import io.camunda.client.api.search.request.TypedFilterableRequest.SearchRequestFilter;
 
 /** Interface for defining Decision Requirements in search queries. */
 public interface DecisionRequirementsFilter extends SearchRequestFilter {
