+++ clients/java/src/main/java/io/camunda/client/api/search/filter/MappingFilter.java
@@ -15,7 +15,7 @@
  */
 package io.camunda.client.api.search.filter;
 
-import io.camunda.client.api.search.request.TypedSearchRequest.SearchRequestFilter;
+import io.camunda.client.api.search.request.TypedFilterableRequest.SearchRequestFilter;
 
 public interface MappingFilter extends SearchRequestFilter {
