+++ clients/java/src/main/java/io/camunda/client/api/search/filter/UserFilter.java
@@ -16,7 +16,7 @@
 package io.camunda.client.api.search.filter;
 
 import io.camunda.client.api.search.filter.builder.StringProperty;
-import io.camunda.client.api.search.request.TypedSearchRequest.SearchRequestFilter;
+import io.camunda.client.api.search.request.TypedFilterableRequest.SearchRequestFilter;
 import java.util.function.Consumer;
 
 public interface UserFilter extends SearchRequestFilter {
